refactor listeners to utils methods
